update js tests
update js tests

+++ .github/scripts/bash/prepare.sh
@@ -61,20 +61,28 @@ then
 fi
 
 #setup php fpm and nginx
-echo -e "${GREEN}setup php-fpm${SET}"
-sudo systemctl enable php$PHP_VERSION-fpm.service
-sudo systemctl start php$PHP_VERSION-fpm.service
-sudo cp -rf  ./.github/artifacts/www.conf /etc/php/$PHP_VERSION/fpm/pool.d/
-sudo systemctl reload php$PHP_VERSION-fpm.service
-sudo systemctl restart php$PHP_VERSION-fpm.service
-sudo systemctl status php$PHP_VERSION-fpm.service
-sudo systemctl enable nginx
-sudo systemctl start nginx
-sudo cp ./.github/artifacts/ui_nginx.conf /etc/nginx/conf.d/
-sudo unlink /etc/nginx/sites-enabled/default
-sudo nginx -t
-sudo systemctl reload nginx
-sudo systemctl restart nginx
+if [ $PIWIK_TEST_TARGET = "Javascript"]
+then
+  echo -e "${GREEN}Setup php -S{SET}"
+  sudo setcap CAP_NET_BIND_SERVICE=+eip $(readlink -f $(which php))
+  tmux new-session -d -s "php-cgi" sudo php -S 127.0.0.1:80
+  tmux ls
+else
+  echo -e "${GREEN}setup php-fpm${SET}"
+  sudo systemctl enable php$PHP_VERSION-fpm.service
+  sudo systemctl start php$PHP_VERSION-fpm.service
+  sudo cp -rf  ./.github/artifacts/www.conf /etc/php/$PHP_VERSION/fpm/pool.d/
+  sudo systemctl reload php$PHP_VERSION-fpm.service
+  sudo systemctl restart php$PHP_VERSION-fpm.service
+  sudo systemctl status php$PHP_VERSION-fpm.service
+  sudo systemctl enable nginx
+  sudo systemctl start nginx
+  sudo cp ./.github/artifacts/ui_nginx.conf /etc/nginx/conf.d/
+  sudo unlink /etc/nginx/sites-enabled/default
+  sudo nginx -t
+  sudo systemctl reload nginx
+  sudo systemctl restart nginx
+fi
 
 echo -e "${GREEN}set folder Permission${SET}"
 cp .github/artifacts/config.ini.github.ui.php config/config.ini.php
